--- a/content/dota_addons/dota_imba/particles/item/silver_edge/imba_silver_edge.vpcf
+++ b/content/dota_addons/dota_imba/particles/item/silver_edge/imba_silver_edge.vpcf
@@ -3,24 +3,25 @@
 	_class = "CParticleSystemDefinition"
 	m_bShouldHitboxesFallbackToRenderBounds = false
 	m_nMaxParticles = 60
-	m_flConstantRadius = 4.000000
-	m_Renderers =
+	m_flConstantRadius = 4.0
+	m_Renderers = 
 	[
 		{
 			_class = "C_OP_RenderRopes"
 			m_nSequenceCombineMode = "SEQUENCE_COMBINE_MODE_USE_SEQUENCE_0"
 			m_bMod2X = true
 			m_hTexture = resource:"materials/particle/beam_blood_03_dark.vtex"
-			m_flRadiusScale = 0.500000
-			m_flFinalTextureScaleU = -1.000000
-			m_flFinalTextureOffsetU = 1.000000
+			m_flRadiusScale = 0.5
+			m_flFinalTextureScaleU = -1.0
+			m_flFinalTextureOffsetU = 1.0
 			m_flTextureVWorldSize = 285.714264
 			m_flTextureVOffset = 0.027902
 			m_flTextureVScrollRate = -214.285706
 			m_nMaxTesselation = 1
+			m_flOverbrightFactor = -1.0
 		},
 	]
-	m_Operators =
+	m_Operators = 
 	[
 		{
 			_class = "C_OP_BasicMovement"
@@ -30,49 +31,39 @@
 		},
 		{
 			_class = "C_OP_InterpolateRadius"
-			m_flBias = 0.600000
-			m_flEndScale = 0.000000
-			m_flStartScale = 12.000000
+			m_flBias = 0.6
+			m_flEndScale = 0.0
+			m_flStartScale = 12.0
 		},
 		{
 			_class = "C_OP_SetControlPointOrientation"
 			m_bUseWorldLocation = true
-			m_vecRotationB =
-			[
-				20.000000,
-				0.000000,
-				0.000000,
-			]
-			m_vecRotation =
-			[
-				20.000000,
-				0.000000,
-				0.000000,
-			]
+			m_vecRotationB = [ 20.0, 0.0, 0.0 ]
+			m_vecRotation = [ 20.0, 0.0, 0.0 ]
 			m_nCP = 0
 		},
 	]
-	m_Initializers =
+	m_Initializers = 
 	[
 		{
 			_class = "C_INIT_RandomLifeTime"
-			m_fLifetimeMax = 0.760000
-			m_fLifetimeMin = 0.760000
+			m_fLifetimeMax = 0.76
+			m_fLifetimeMin = 0.76
 		},
 		{
 			_class = "C_INIT_RingWave"
-			m_flRoll = 180.000000
-			m_flInitialRadius = 110.000000
-			m_flPitch = -90.000000
+			m_flRoll = 180.0
+			m_flInitialRadius = 110.0
+			m_flPitch = -90.0
 			m_bEvenDistribution = true
-			m_flParticlesPerOrbit = 155.000000
+			m_flParticlesPerOrbit = 155.0
 		},
 		{
 			_class = "C_INIT_RemapScalar"
-			m_flInputMin = 0.065000
-			m_flInputMax = 0.075000
-			m_flOutputMin = 1.000000
-			m_flOutputMax = 0.000000
+			m_flInputMin = 0.065
+			m_flInputMax = 0.075
+			m_flOutputMin = 1.0
+			m_flOutputMax = 0.0
 			m_bScaleInitialRange = true
 		},
 		{
@@ -83,38 +74,43 @@
 		},
 		{
 			_class = "C_INIT_PositionOffset"
-			m_OffsetMax =
-			[
-				0.000000,
-				0.000000,
-				80.000000,
-			]
-			m_OffsetMin =
-			[
-				0.000000,
-				0.000000,
-				80.000000,
-			]
+			m_OffsetMax = [ 0.0, 0.0, 80.0 ]
+			m_OffsetMin = [ 0.0, 0.0, 80.0 ]
 		},
 	]
-	m_Emitters =
+	m_Emitters = 
 	[
 		{
 			_class = "C_OP_ContinuousEmitter"
-			m_flEmissionDuration = 0.075000
-			m_flEmitRate = 805.000000
+			m_flEmissionDuration = 0.075
+			m_flEmitRate = 805.0
 		},
 	]
-	m_Children =
+	m_Children = 
 	[
 		{
-			m_ChildRef = resource:"particles/items3_fx/silver_edge_d.vpcf"
+			m_ChildRef = resource:"particles/item/silver_edge/imba_silver_edge_a.vpcf"
 		},
 		{
-			m_ChildRef = resource:"particles/items3_fx/silver_edge_b.vpcf"
+			m_ChildRef = resource:"particles/item/silver_edge/imba_silver_edge_b.vpcf"
 		},
 		{
-			m_ChildRef = resource:"particles/items3_fx/silver_edge_c.vpcf"
+			m_ChildRef = resource:"particles/item/silver_edge/imba_silver_edge_c.vpcf"
+		},
+	]
+	m_controlPointConfigurations = 
+	[
+		{
+			m_name = "preview"
+			m_drivers = 
+			[
+				{
+					m_iAttachType = "PATTACH_WORLDORIGIN"
+					m_vecOffset = [ 0.0, 0.0, 0.0 ]
+					m_angOffset = [ null, null, null ]
+					m_entityName = "self"
+				},
+			]
 		},
 	]
 }