--- a/content/dota_addons/dota_imba/particles/item/silver_edge/silver_edge_shadow_rip.vpcf
+++ b/content/dota_addons/dota_imba/particles/item/silver_edge/silver_edge_shadow_rip.vpcf
@@ -1,0 +1,480 @@
+<!-- kv3 encoding:text:version{e21c7f3c-8a33-41c5-9977-a76d3a32aa0d} format:generic:version{7412167c-06e9-4698-aff2-e63eb59037e7} -->
+{
+	_class = "CParticleSystemDefinition"
+	m_bShouldHitboxesFallbackToRenderBounds = false
+	m_nMaxParticles = 300
+	m_flConstantRadius = 64.0
+	m_bShouldSort = false
+	m_Renderers = 
+	[
+		{
+			_class = "C_OP_RenderRopes"
+			m_nSequenceCombineMode = "SEQUENCE_COMBINE_MODE_USE_SEQUENCE_0"
+			m_bAdditive = true
+			m_flOverbrightFactor = -40.0
+			m_hTexture = resource:"materials/particle/beam_fire_01.vtex"
+			m_nScaleCP1 = 0
+			m_nScaleCP2 = 1
+			m_flFinalTextureScaleU = -1.0
+			m_flFinalTextureOffsetU = 1.0
+			m_flTextureVWorldSize = 750.0
+			m_flTextureVOffset = 250.0
+			m_flTextureVScrollRate = -150.0
+			m_nMaxTesselation = 2
+			m_nMinTesselation = 2
+			m_flFeatheringMaxDist = 32.0
+			VisibilityInputs = 
+			{
+				m_flCameraBias = 64.0
+			}
+			m_nOrientationType = 2
+			m_flAddSelfAmount = 2.0
+			m_bMod2X = true
+			m_bLightenMode = true
+			m_flDepthBias = 24.0
+		},
+		{
+			_class = "C_OP_RenderRopes"
+			m_nMaxTesselation = 4
+			m_flTextureVWorldSize = 1000.0
+			m_flRadiusScale = 2.0
+			m_bAdditive = true
+			m_flOverbrightFactor = 10.0
+			m_hTexture = resource:"materials/particle/beam_hotblue.vtex"
+			m_flFeatheringMaxDist = 32.0
+			VisibilityInputs = 
+			{
+				m_flCameraBias = 32.0
+			}
+			m_bDisableOperator = true
+		},
+		{
+			_class = "C_OP_RenderRopes"
+			m_flTextureVWorldSize = 800.0
+			m_flTextureVScrollRate = 200.0
+			m_nOrientationType = 3
+			m_flOverbrightFactor = -4.0
+			m_hTexture = resource:"materials/particle/beam_fire_01.vtex"
+			m_flFeatheringMaxDist = 32.0
+			VisibilityInputs = 
+			{
+				m_flCameraBias = 32.0
+			}
+			m_bMod2X = true
+		},
+		{
+			_class = "C_OP_RenderScreenShake"
+			m_flDurationScale = 0.125
+			m_flFrequencyScale = 2.75
+			m_flAmplitudeScale = 1.25
+			m_nAmplitudeField = 23
+			m_flRadiusScale = 100.0
+			m_nRadiusField = 23
+			m_nDurationField = 23
+			m_nFrequencyField = 23
+		},
+	]
+	m_Operators = 
+	[
+		{
+			_class = "C_OP_Decay"
+		},
+		{
+			_class = "C_OP_BasicMovement"
+			m_fDrag = 0.26
+		},
+		{
+			_class = "C_OP_InterpolateRadius"
+			m_flEndScale = 7.0
+			m_flBias = 0.9
+			m_flStartScale = 0.0
+		},
+		{
+			_class = "C_OP_ControlpointLight"
+			m_LightZeroDist4 = 100.0
+			m_LightFiftyDist4 = 64.0
+			m_LightColor4 = [ 75, 0, 130, 255 ]
+			m_nControlPoint4 = 4
+			m_LightZeroDist3 = 100.0
+			m_LightFiftyDist3 = 64.0
+			m_LightColor3 = [ 75, 0, 130, 255 ]
+			m_nControlPoint3 = 3
+			m_LightZeroDist2 = 100.0
+			m_LightFiftyDist2 = 64.0
+			m_LightColor2 = [ 75, 0, 130, 255 ]
+			m_nControlPoint2 = 2
+			m_LightZeroDist1 = 100.0
+			m_LightFiftyDist1 = 64.0
+			m_LightColor1 = [ 75, 0, 130, 255 ]
+			m_nControlPoint1 = 1
+			m_bUseHLambert = false
+			m_bClampLowerRange = true
+		},
+		{
+			_class = "C_OP_ControlpointLight"
+			m_bClampLowerRange = true
+			m_nControlPoint1 = 5
+			m_LightFiftyDist1 = 64.0
+			m_LightZeroDist1 = 100.0
+			m_nControlPoint2 = 6
+			m_LightFiftyDist2 = 64.0
+			m_LightZeroDist2 = 100.0
+			m_nControlPoint3 = 7
+			m_LightFiftyDist3 = 64.0
+			m_LightZeroDist3 = 100.0
+			m_nControlPoint4 = 8
+			m_LightFiftyDist4 = 64.0
+			m_LightZeroDist4 = 100.0
+			m_bUseHLambert = false
+		},
+		{
+			_class = "C_OP_ControlpointLight"
+			m_bClampLowerRange = true
+			m_nControlPoint1 = 9
+			m_LightColor1 = [ 139, 0, 139, 255 ]
+			m_LightFiftyDist1 = 64.0
+			m_LightZeroDist1 = 100.0
+			m_nControlPoint2 = 10
+			m_LightColor2 = [ 139, 0, 139, 255 ]
+			m_LightFiftyDist2 = 64.0
+			m_LightZeroDist2 = 100.0
+			m_nControlPoint3 = 11
+			m_LightColor3 = [ 75, 0, 130, 255 ]
+			m_LightFiftyDist3 = 64.0
+			m_LightZeroDist3 = 100.0
+			m_nControlPoint4 = 12
+			m_LightColor4 = [ 75, 0, 130, 255 ]
+			m_LightFiftyDist4 = 64.0
+			m_LightZeroDist4 = 100.0
+			m_bUseHLambert = false
+		},
+		{
+			_class = "C_OP_ControlpointLight"
+			m_LightZeroDist4 = 100.0
+			m_LightFiftyDist4 = 64.0
+			m_LightColor4 = [ 73, 110, 255, 255 ]
+			m_nControlPoint4 = 16
+			m_LightZeroDist3 = 100.0
+			m_LightFiftyDist3 = 64.0
+			m_LightColor3 = [ 73, 110, 255, 255 ]
+			m_nControlPoint3 = 15
+			m_LightZeroDist2 = 100.0
+			m_LightFiftyDist2 = 64.0
+			m_LightColor2 = [ 73, 110, 255, 255 ]
+			m_nControlPoint2 = 14
+			m_LightZeroDist1 = 100.0
+			m_LightFiftyDist1 = 64.0
+			m_LightColor1 = [ 73, 110, 255, 255 ]
+			m_nControlPoint1 = 13
+			m_bClampLowerRange = true
+			m_bUseHLambert = false
+			m_bDisableOperator = true
+		},
+		{
+			_class = "C_OP_SetControlPointPositions"
+			m_nCP1 = 13
+			m_vecCP1Pos = [ 0.0, 0.0, -512.0 ]
+			m_nCP2 = 14
+			m_vecCP2Pos = [ 0.0, 0.0, -512.0 ]
+			m_nCP3 = 15
+			m_vecCP3Pos = [ 0.0, 0.0, -512.0 ]
+			m_nCP4 = 16
+			m_vecCP4Pos = [ 0.0, 0.0, -512.0 ]
+			m_bSetOnce = true
+		},
+		{
+			_class = "C_OP_SetControlPointPositions"
+			m_nCP1 = 12
+			m_vecCP1Pos = [ 0.0, 0.0, -512.0 ]
+			m_nCP2 = 11
+			m_vecCP2Pos = [ 0.0, 0.0, -512.0 ]
+			m_nCP3 = 10
+			m_vecCP3Pos = [ 0.0, 0.0, -512.0 ]
+			m_nCP4 = 9
+			m_vecCP4Pos = [ 0.0, 0.0, -512.0 ]
+			m_bSetOnce = true
+		},
+		{
+			_class = "C_OP_SetControlPointPositions"
+			m_nCP1 = 8
+			m_vecCP1Pos = [ 0.0, 0.0, -512.0 ]
+			m_nCP2 = 7
+			m_vecCP2Pos = [ 0.0, 0.0, -512.0 ]
+			m_nCP3 = 6
+			m_vecCP3Pos = [ 0.0, 0.0, -512.0 ]
+			m_nCP4 = 5
+			m_vecCP4Pos = [ 0.0, 0.0, -512.0 ]
+			m_bSetOnce = true
+		},
+		{
+			_class = "C_OP_SetControlPointPositions"
+			m_bSetOnce = true
+			m_vecCP4Pos = [ 0.0, 0.0, -512.0 ]
+			m_vecCP3Pos = [ 0.0, 0.0, -512.0 ]
+			m_vecCP2Pos = [ 0.0, 0.0, -512.0 ]
+			m_vecCP1Pos = [ 0.0, 0.0, -512.0 ]
+		},
+		{
+			_class = "C_OP_MovementPlaceOnGround"
+			m_flOffset = 24.0
+			m_flMaxTraceLength = 2048.0
+			m_flTolerance = 0.0
+			m_flTraceOffset = 512.0
+			m_CollisionGroupName = "DEBRIS"
+			m_bIncludeWater = true
+		},
+		{
+			_class = "C_OP_FadeOutSimple"
+			m_flFadeOutTime = 0.5
+		},
+		{
+			_class = "C_OP_FadeInSimple"
+			m_flFadeInTime = 0.4
+		},
+		{
+			_class = "C_OP_ColorInterpolateRandom"
+			m_ColorFadeMin = [ 175, 250, 120, 255 ]
+			m_ColorFadeMax = [ 175, 250, 120, 255 ]
+			m_flFadeEndTime = 0.5
+		},
+	]
+	m_Initializers = 
+	[
+		{
+			_class = "C_INIT_RandomLifeTime"
+			m_fLifetimeMin = 0.9
+			m_fLifetimeMax = 0.9
+		},
+		{
+			_class = "C_INIT_RandomAlpha"
+			m_bDisableOperator = true
+			m_nAlphaMin = 0
+		},
+		{
+			_class = "C_INIT_RingWave"
+			m_flInitialRadius = 1.0
+			m_bEvenDistribution = true
+			m_flParticlesPerOrbit = 200.0
+			m_flInitialSpeedMin = 5000.0
+			m_flInitialSpeedMax = 5000.0
+			m_flYaw = 315.0
+		},
+		{
+			_class = "C_INIT_PositionOffset"
+			m_OffsetMax = [ 0.0, 0.0, 32.0 ]
+			m_OffsetMin = [ 0.0, 0.0, 32.0 ]
+		},
+		{
+			_class = "C_INIT_RemapParticleCountToScalar"
+			m_nFieldOutput = 16
+			m_nInputMin = 1
+		},
+		{
+			_class = "C_INIT_RemapParticleCountToScalar"
+			m_nFieldOutput = 16
+			m_nInputMin = 38
+			m_nInputMax = 47
+			m_flOutputMin = 1.0
+			m_flOutputMax = 0.0
+			m_bScaleInitialRange = true
+		},
+		{
+			_class = "C_INIT_RemapParticleCountToScalar"
+			m_nFieldOutput = 18
+			m_nInputMax = 20
+		},
+		{
+			_class = "C_INIT_RemapParticleCountToScalar"
+			m_nFieldOutput = 18
+			m_nInputMax = 20
+			m_flOutputMin = 1.0
+			m_flOutputMax = 0.0
+			m_bActiveRange = true
+			m_bInvert = true
+		},
+		{
+			_class = "C_INIT_RemapScalarToVector"
+			m_nFieldInput = 18
+			m_nFieldOutput = 6
+			m_vecOutputMin = [ 0.5, 0.41, 0.032 ]
+			m_vecOutputMax = [ 0.199, 0.26, 0.414 ]
+			m_bLocalCoords = false
+		},
+		{
+			_class = "C_INIT_CreationNoise"
+			m_nFieldOutput = 7
+			m_flOutputMin = 0.5
+			m_flNoiseScaleLoc = 50.0
+		},
+		{
+			_class = "C_INIT_RemapScalar"
+			m_nFieldInput = 18
+			m_nFieldOutput = 16
+			m_flOutputMax = 0.7
+		},
+	]
+	m_Emitters = 
+	[
+		{
+			_class = "C_OP_InstantaneousEmitter"
+			m_nParticlesToEmit = 48
+		},
+	]
+	m_ForceGenerators = 
+	[
+		{
+			_class = "C_OP_AttractToControlPoint"
+			m_fForceAmount = -150.0
+			m_fFalloffPower = -0.1
+		},
+	]
+	m_ConstantColor = [ 75, 0, 130, 255 ]
+	m_controlPointConfigurations = 
+	[
+		{
+			m_name = "preview"
+			m_drivers = 
+			[
+				{
+					m_iAttachType = "PATTACH_WORLDORIGIN"
+					m_vecOffset = [ 0.0, 0.0, 0.0 ]
+					m_angOffset = [ null, null, null ]
+					m_entityName = "self"
+				},
+				{
+					m_iControlPoint = 1
+					m_iAttachType = "PATTACH_WORLDORIGIN"
+					m_vecOffset = [ 0.0, 0.0, 0.0 ]
+					m_angOffset = [ null, null, null ]
+					m_entityName = "self"
+				},
+				{
+					m_iControlPoint = 2
+					m_iAttachType = "PATTACH_WORLDORIGIN"
+					m_vecOffset = [ 0.0, 0.0, 0.0 ]
+					m_angOffset = [ null, null, null ]
+					m_entityName = "self"
+				},
+				{
+					m_iControlPoint = 3
+					m_iAttachType = "PATTACH_WORLDORIGIN"
+					m_vecOffset = [ 0.0, 0.0, 0.0 ]
+					m_angOffset = [ null, null, null ]
+					m_entityName = "self"
+				},
+				{
+					m_iControlPoint = 4
+					m_iAttachType = "PATTACH_WORLDORIGIN"
+					m_vecOffset = [ 0.0, 0.0, 0.0 ]
+					m_angOffset = [ null, null, null ]
+					m_entityName = "self"
+				},
+				{
+					m_iControlPoint = 5
+					m_iAttachType = "PATTACH_WORLDORIGIN"
+					m_vecOffset = [ 0.0, 0.0, 0.0 ]
+					m_angOffset = [ null, null, null ]
+					m_entityName = "self"
+				},
+				{
+					m_iControlPoint = 6
+					m_iAttachType = "PATTACH_WORLDORIGIN"
+					m_vecOffset = [ 0.0, 0.0, 0.0 ]
+					m_angOffset = [ null, null, null ]
+					m_entityName = "self"
+				},
+				{
+					m_iControlPoint = 7
+					m_iAttachType = "PATTACH_WORLDORIGIN"
+					m_vecOffset = [ 0.0, 0.0, 0.0 ]
+					m_angOffset = [ null, null, null ]
+					m_entityName = "self"
+				},
+				{
+					m_iControlPoint = 8
+					m_iAttachType = "PATTACH_WORLDORIGIN"
+					m_vecOffset = [ 0.0, 0.0, 0.0 ]
+					m_angOffset = [ null, null, null ]
+					m_entityName = "self"
+				},
+				{
+					m_iControlPoint = 9
+					m_iAttachType = "PATTACH_WORLDORIGIN"
+					m_vecOffset = [ 0.0, 0.0, 0.0 ]
+					m_angOffset = [ null, null, null ]
+					m_entityName = "self"
+				},
+				{
+					m_iControlPoint = 10
+					m_iAttachType = "PATTACH_WORLDORIGIN"
+					m_vecOffset = [ 0.0, 0.0, 0.0 ]
+					m_angOffset = [ null, null, null ]
+					m_entityName = "self"
+				},
+				{
+					m_iControlPoint = 11
+					m_iAttachType = "PATTACH_WORLDORIGIN"
+					m_vecOffset = [ 0.0, 0.0, 0.0 ]
+					m_angOffset = [ null, null, null ]
+					m_entityName = "self"
+				},
+				{
+					m_iControlPoint = 12
+					m_iAttachType = "PATTACH_WORLDORIGIN"
+					m_vecOffset = [ 0.0, 0.0, 0.0 ]
+					m_angOffset = [ null, null, null ]
+					m_entityName = "self"
+				},
+				{
+					m_iControlPoint = 13
+					m_iAttachType = "PATTACH_WORLDORIGIN"
+					m_vecOffset = [ 0.0, 0.0, 0.0 ]
+					m_angOffset = [ null, null, null ]
+					m_entityName = "self"
+				},
+				{
+					m_iControlPoint = 14
+					m_iAttachType = "PATTACH_WORLDORIGIN"
+					m_vecOffset = [ 0.0, 0.0, 0.0 ]
+					m_angOffset = [ null, null, null ]
+					m_entityName = "self"
+				},
+				{
+					m_iControlPoint = 15
+					m_iAttachType = "PATTACH_WORLDORIGIN"
+					m_vecOffset = [ 0.0, 0.0, 0.0 ]
+					m_angOffset = [ null, null, null ]
+					m_entityName = "self"
+				},
+				{
+					m_iControlPoint = 16
+					m_iAttachType = "PATTACH_WORLDORIGIN"
+					m_vecOffset = [ 0.0, 0.0, 0.0 ]
+					m_angOffset = [ null, null, null ]
+					m_entityName = "self"
+				},
+			]
+		},
+	]
+	m_Children = 
+	[
+		{
+			m_ChildRef = resource:"particles/item/silver_edge/silver_edge_shadow_rip_a.vpcf"
+		},
+		{
+			m_ChildRef = resource:"particles/item/silver_edge/silver_edge_shadow_rip_a1.vpcf"
+			m_flDelay = 0.25
+		},
+		{
+			m_ChildRef = resource:"particles/item/silver_edge/silver_edge_shadow_rip_a2.vpcf"
+			m_flDelay = 0.3
+		},
+		{
+			m_ChildRef = resource:"particles/item/silver_edge/silver_edge_shadow_rip_b.vpcf"
+			m_bDisableChild = true
+		},
+		{
+			m_ChildRef = resource:"particles/item/silver_edge/silver_edge_shadow_rip_c.vpcf"
+		},
+	]
+}