… content/dota_addons/dota_imba/particles/item/silver_edge/silver_edge_break.vpcf
… content/dota_addons/dota_imba/particles/item/silver_edge/silver_edge_break_caster.vpcf
… content/dota_addons/dota_imba/particles/item/silver_edge/silver_edge_break_dark.vpcf
… content/dota_addons/dota_imba/particles/item/silver_edge/silver_edge_break_glow_caster.vpcf
… content/dota_addons/dota_imba/particles/item/silver_edge/silver_edge_break_ring.vpcf
… content/dota_addons/dota_imba/particles/item/silver_edge/silver_edge_shadow_rip_a.vpcf
… content/dota_addons/dota_imba/particles/item/silver_edge/silver_edge_shadow_rip_a0.vpcf
… content/dota_addons/dota_imba/particles/item/silver_edge/silver_edge_shadow_rip_a1.vpcf
… content/dota_addons/dota_imba/particles/item/silver_edge/silver_edge_shadow_rip_a2.vpcf
… content/dota_addons/dota_imba/particles/item/silver_edge/silver_edge_shadow_rip_b.vpcf
… content/dota_addons/dota_imba/particles/item/silver_edge/silver_edge_shadow_rip_c.vpcf
… content/dota_addons/dota_imba/particles/item/silver_edge/silver_edge_target.vpcf
… content/dota_addons/dota_imba/soundevents/imba_item_soundevents.vsndevts
… content/dota_addons/dota_imba/sounds/item/silver_edge/silver_edge_shadow_rip.mp3
… game/dota_addons/dota_imba/scripts/npc/npc_items_custom.txt
… game/dota_addons/dota_imba/scripts/vscripts/internal/funcs.lua
… game/dota_addons/dota_imba/scripts/vscripts/items/item_blink.lua
… game/dota_addons/dota_imba/scripts/vscripts/items/item_shadow_blade.lua
… game/dota_addons/dota_imba/scripts/vscripts/items/item_silver_edge.lua