**Added 3 new functions:**
IsHeroDamage
CalculateDirection
FindUnitsInCone

Sorted the sound event file for items from A to Z.

# **Reworked Shadow Blade**
(stats are the same)
## OLD: (datadriven)
Can be casted to grant invisibility, flying movement and 300 night vision. Flying movement is disabled for 2 seconds upon taking hero-based damage.
Upon hitting to break invisiblity, that attack will deal 175 additional damage.

## NEW:(lua)
### Active: _Shadow Walk_
Makes you invisible for 14 seconds, or until you attack or cast a spell.  While Shadow Walk is active, you move 20% faster and can move through units **and terrain**.  Flying movement is disabled for 2 seconds upon taking hero-based damage.
If attacking to end the invisibility, you gain 175 bonus damage on that attack,  **and apply _Shadowy Panic_ to the target.**

### Passive: _Shadowy Panic_ 
 Attacking to break invisiblity cripples the target's senses, slowing their turn rate by 80% for 2 seconds.


# **Reworked Silver Edge**
## OLD: (datadriven)
Can be casted to grant invisibility, flying movement and 600 night vision. Flying movement is disabled for 2 seconds upon taking hero-based damage.
Upon hitting to break invisiblity, that attack will deal 175 additional damage and apply break to the target, reducing its damage by 50% and disabling its passives for 5 seconds.

## NEW:(lua)
### Active: _Shadow Walk_
Makes you invisible for 16 seconds, or until you attack or cast a spell.  While Shadow Walk is active, you move 20% faster and can move through units **and terrain.**  Flying movement is disabled for 1 second upon taking hero-based damage.
If attacking to end the invisibility, you break the target, disabling their passives for 5 seconds, and reducing their damage output by 50%. **Additionally, you _Shadow Rip_.** 
### Passive: _Shadow Rip_
Attacking to end Invisibility rips the air apart, releasing panic-inducing clouds of darkness, that engulf the target and enemies behind him. _Shadow Rip_ deals 250 damage, disables enemies' passives, reduces their damage output by 50%, and slows their turn rate by 80% for 2 seconds.